+++ packages/cli/src/commands/service/dev.ts
@@ -1,23 +1,22 @@
 import { Command, flags } from '@oclif/command'
 import Listr from 'listr'
 import LCD from '@mesg/api/lib/lcd'
-import API from '@mesg/api'
+import Orchestrator from '@mesg/orchestrator'
+import Event from '@mesg/orchestrator/lib/typedef/event'
+import Execution from '@mesg/orchestrator/lib/typedef/execution'
+import * as grpc from 'grpc'
 import chalk from 'chalk'
 import { decode } from '@mesg/api/lib/util/encoder'
 import { parseLog, listContainers } from '../../utils/docker'
 import * as Environment from '../../utils/environment-tasks'
 import * as Service from '../../utils/service'
 import * as Runner from '../../utils/runner'
-import * as base58 from "@mesg/api/lib/util/base58";
-import { ExecutionStatus } from "@mesg/api/lib/types";
+import { Status } from "@mesg/orchestrator/lib/execution";
 import version from '../../version'
 import { IService, IDefinition } from '@mesg/api/lib/service-lcd'
-import { IRunner } from '@mesg/api/lib/runner-lcd'
 import { Stream } from 'stream'
-import { IEvent } from "@mesg/api/lib/event";
-import { IExecution } from "@mesg/api/lib/execution";
-import { Stream as GRPCStream } from "@mesg/api/lib/util/grpc";
 import { RunnerInfo } from '@mesg/runner'
+import sign from '../../utils/sign'
 
 const ipfsClient = require('ipfs-http-client')
 
@@ -41,13 +40,14 @@ export default class Dev extends Command {
   }]
 
   private lcdEndpoint = 'http://localhost:1317'
+  private orchestratorEndpoint = 'localhost:50052'
   private lcd = new LCD(this.lcdEndpoint)
-  private grpc = new API('localhost:50052')
+  private orchestrator = new Orchestrator(this.orchestratorEndpoint)
   private ipfsClient = ipfsClient('ipfs.app.mesg.com', '5001', { protocol: 'http' })
 
   private logs: Stream[]
-  private events: GRPCStream<IEvent>
-  private results: GRPCStream<IExecution>
+  private events: grpc.ClientReadableStream<Event.mesg.types.IEvent>
+  private results: grpc.ClientReadableStream<Execution.mesg.types.IExecution>
 
   async run() {
     const { args, flags } = this.parse(Dev)
@@ -67,13 +67,13 @@ export default class Dev extends Command {
       {
         title: 'Creating service',
         task: async ctx => {
-          service = await Service.create(this.lcd, definition, ctx.mnemonic)
+          service = await Service.create(this.lcd, definition, ctx.config.mnemonic)
         }
       },
       {
         title: 'Starting service',
         task: async ctx => {
-          runner = await Runner.create(this.lcdEndpoint, ctx.mnemonic, service.hash, flags.env)
+          runner = await Runner.create(this.lcdEndpoint, this.orchestratorEndpoint, ctx.config.mnemonic, ctx.engineAddress, service.hash, flags.env)
         }
       },
       {
@@ -93,32 +93,34 @@ export default class Dev extends Command {
           },
           {
             title: 'Fetching events\' logs',
-            task: () => {
-              this.events = this.grpc.event.stream({
+            task: ctx => {
+              const payload = {
                 filter: {
-                  instanceHash: base58.decode(runner.instanceHash)
+                  instanceHash: runner.instanceHash
                 }
-              })
+              }
+              this.events = this.orchestrator.event.stream(payload, sign(payload, ctx.config.mnemonic))
             }
           },
           {
             title: 'Fetching executions\' logs',
-            task: async () => {
-              this.results = this.grpc.execution.stream({
+            task: async ctx => {
+              const payload = {
                 filter: {
-                  executorHash: base58.decode(runner.hash),
+                  executorHash: runner.hash,
                   statuses: [
-                    ExecutionStatus.COMPLETED,
-                    ExecutionStatus.FAILED,
+                    Status.Completed,
+                    Status.Failed,
                   ]
                 }
-              })
+              }
+              this.results = this.orchestrator.execution.stream(payload, sign(payload, ctx.config.mnemonic))
             }
           }
         ])
       }
     ])
-    const res = await tasks.run({
+    const { config, engineAddress } = await tasks.run({
       configDir: this.config.dataDir,
       pull: flags.pull,
       version: flags.version,
@@ -156,7 +158,7 @@ export default class Dev extends Command {
           title: 'Stopping service',
           skip: () => !service && !runner,
           task: async () => {
-            return Runner.stop(this.lcdEndpoint, res.mnemonic, runner.hash)
+            return Runner.stop(this.lcdEndpoint, this.orchestratorEndpoint, config.mnemonic, engineAddress, runner.hash)
           }
         },
         Environment.stop,