+++ packages/cli/src/commands/service/execute.ts
@@ -1,18 +1,23 @@
 import { flags, Command } from '@oclif/command'
 import { readFileSync } from 'fs'
-import { cli } from 'cli-ux'
 import LCD from '@mesg/api/lib/lcd'
-import Application from '@mesg/application'
-import API from '@mesg/api'
+import * as grpc from 'grpc'
 import { decode } from '@mesg/api/lib/util/encoder'
 import Listr from 'listr'
-import * as base58 from "@mesg/api/lib/util/base58";
 import ServiceType from "@mesg/api/lib/typedef/service";
 import ExecutionType from "@mesg/api/lib/typedef/execution";
 import { resolveSIDRunner } from "@mesg/api/lib/util/resolve";
 import { IRunner } from '@mesg/api/lib/runner-lcd'
 import { IExecution } from '@mesg/api/lib/execution'
 import { convert } from '../../utils/input'
+import Orchestrator from '@mesg/orchestrator'
+import * as Type from '@mesg/orchestrator/lib/typedef/execution'
+import { CreateRequest, StreamRequest, Status } from '@mesg/orchestrator/lib/execution'
+import sign from '../../utils/sign'
+import uuid from 'uuid'
+import { generateConfig } from '../../utils/config'
+import { toStruct } from '@mesg/api/lib/struct'
+import styledJSON from 'cli-ux/lib/styled/json'
 
 export default class Execute extends Command {
   static description = 'Execute a task on a running service'
@@ -41,15 +46,20 @@ export default class Execute extends Command {
   }]
 
   private lcd = new LCD(`http://localhost:1317`)
-  private app = new Application(new API(`localhost:50052`))
+  private orchestrator = new Orchestrator(`localhost:50052`)
 
   async run() {
     const { args, flags } = this.parse(Execute)
 
+    const id = uuid()
+
     let runner: IRunner
     let task: ServiceType.mesg.types.Service.ITask
     let inputs: ExecutionType.mesg.protobuf.IStruct
+    let logs: grpc.ClientReadableStream<Type.mesg.types.IExecution>
     let execution: IExecution
+
+    const config = generateConfig(this.config.dataDir)
 
     const tasks = new Listr([
       {
@@ -79,20 +89,46 @@ export default class Execute extends Command {
         }
       },
       {
+        title: 'Listening for result',
+        task: async () => {
+          const payload: StreamRequest = {
+            filter: {
+              executorHash: runner.hash,
+              instanceHash: runner.instanceHash,
+              tags: [id],
+              taskKey: args.TASK,
+              statuses: [Status.Completed, Status.Failed]
+            }
+          }
+          logs = await this.orchestrator.execution.stream(payload, sign(payload, config.mnemonic))
+        }
+      },
+      {
         title: 'Executing task',
         task: async () => {
-          execution = await this.app.executeTaskAndWaitResult({
-            eventHash: base58.decode(flags.eventHash || '6aUPZhmnFKiSsHXRaddbnqsKKi9KogbQNiKUcpivaohb'),
-            executorHash: base58.decode(runner.hash),
-            inputs,
-            tags: ['CLI'],
-            taskKey: args.TASK
+          const payload: CreateRequest = {
+            executorHash: runner.hash,
+            inputs: inputs,
+            taskKey: args.TASK,
+            price: '10000atto',
+            tags: [id]
+          }
+          await this.orchestrator.execution.create(
+            payload,
+            sign({ ...payload, inputs: toStruct(payload.inputs) }, config.mnemonic)
+          )
+          return new Promise((resolve, reject) => {
+            logs.once('data', (exec: IExecution) => {
+              execution = exec
+              execution.status === Status.Completed ? resolve(execution) : reject(new Error(execution.error))
+            })
           })
         }
       }
     ])
     await tasks.run()
 
-    cli.styledJSON(decode(execution.outputs))
+    styledJSON(decode(execution.outputs))
+    process.exit(0)
   }
 }