+++ packages/runner/src/index.ts
@@ -1,5 +1,7 @@
 import API from '@mesg/api/lib/lcd'
+import Orchestrator from '@mesg/orchestrator'
 import Account from "@mesg/api/lib/account-lcd"
+import sortObject from '@mesg/api/lib/util/sort-object'
 import { IService } from '@mesg/api/lib/service-lcd'
 import Transaction from '@mesg/api/lib/transaction'
 
@@ -8,29 +10,43 @@ export type RunnerInfo = {
   instanceHash: string
 }
 
+export type Env = { [key: string]: string }
+
 export interface Provider {
-  start(service: IService, env: string[], runnerHash: string, instanceHash: string, token: string): Promise<boolean>
-  stop(runnerHash: string): Promise<void>
+  start(service: IService, env: Env, runnerHash: string): Promise<boolean>
+  stop(runnerHash: string, serviceHash: string): Promise<void>
 }
 
 export default class Runner {
 
   private _provider: Provider
   private _api: API
+  private _orchestrator: Orchestrator
   private _mnemonic: string
+  private _engineAddress: string
 
-  constructor(provider: Provider, endpoint: string, mnemonic: string) {
-    this._api = new API(endpoint)
+  constructor(provider: Provider, lcdEndpoint: string, orchestratorEndpoint: string, mnemonic: string, engineAddress: string) {
+    this._api = new API(lcdEndpoint)
+    this._orchestrator = new Orchestrator(orchestratorEndpoint)
     this._provider = provider
     this._mnemonic = mnemonic
+    this._engineAddress = engineAddress
   }
 
   async start(serviceHash: string, env: string[]): Promise<RunnerInfo> {
-    const account = await this._api.account.import(this._mnemonic)
     const service = await this._api.service.get(serviceHash)
-    const { runnerHash, instanceHash, envHash } = await this._api.runner.hash(account.address, serviceHash, env)
-    const token = this.createRegisterRunner(serviceHash, envHash)
-    await this._provider.start(service, env, runnerHash, instanceHash, token)
+    const { runnerHash, instanceHash, envHash } = await this._api.runner.hash(this._engineAddress, serviceHash, env)
+    const envObj = (env || []).reduce((prev, x) => ({ ...prev, [x.split('=')[0]]: x.split('=')[1] }), {})
+    await this._provider.start(service, {
+      ...envObj,
+      MESG_ENDPOINT: `engine:50052`,
+      MESG_SERVICE_HASH: service.hash,
+      MESG_ENV_HASH: envHash,
+      MESG_REGISTER_SIGNATURE: this.sign({
+        serviceHash: serviceHash,
+        envHash: envHash
+      })
+    }, runnerHash)
     return {
       instanceHash,
       hash: runnerHash
@@ -38,22 +54,15 @@ export default class Runner {
   }
 
   async stop(runnerHash: string): Promise<void> {
-    const account = await this._api.account.import(this._mnemonic)
-    const tx = await this._api.createTransaction(
-      [this._api.runner.deleteMsg(account.address, runnerHash)],
-      account
-    )
-    await this._api.broadcast(tx.signWithMnemonic(this._mnemonic), "block")
-    await this._provider.stop(runnerHash)
+    const runner = await this._api.runner.get(runnerHash)
+    const instance = await this._api.instance.get(runner.instanceHash)
+    const payload = { runnerHash }
+    await this._orchestrator.runner.delete(payload, this.sign(payload))
+    await this._provider.stop(runnerHash, instance.serviceHash)
   }
 
-  private createRegisterRunner(serviceHash: string, envHash: string): string {
-    const value = {
-      serviceHash: serviceHash,
-      envHash: envHash
-    }
-    const ecdsa = Transaction.sign(JSON.stringify(value), Account.getPrivateKey(this._mnemonic))
-    const signature = Buffer.from(ecdsa.signature).toString('base64')
-    return JSON.stringify({ signature, value })
+  private sign(data: Object): string {
+    const ecdsa = Transaction.sign(JSON.stringify(sortObject(data)), Account.getPrivateKey(this._mnemonic))
+    return Buffer.from(ecdsa.signature).toString('base64')
   }
 }