+++ packages/service/src/index.ts
@@ -3,35 +3,44 @@ import * as fs from 'fs'
 import * as grpc from 'grpc'
 import * as protoLoader from '@grpc/proto-loader'
 import * as path from 'path'
+import Runner from '@mesg/orchestrator/lib/runner'
 import { decode, encode } from '@mesg/api/lib/util/encoder'
+import * as base58 from '@mesg/api/lib/util/base58'
 import { IExecution } from '@mesg/api/lib/execution';
 import { EventCreateOutputs } from '@mesg/api/lib/event';
 import { EventEmitter } from 'events'
 
 type Options = {
   endpoint?: string
-  payload?: string
+  signature?: string
   definition?: any
 }
 
 class Service {
   private definition: any
   private tasks: Tasks
-  private _client: any
-  private credential: Promise<grpc.Metadata>
+  private _client: Promise<any>
+  private _token: grpc.Metadata
 
   constructor(options: Options = {}) {
     this.definition = options.definition || YAML.safeLoad(fs.readFileSync('./mesg.yml').toString());
-    const protoDefinition = grpc.loadPackageDefinition(protoLoader.loadSync(path.join(__dirname, 'runner.proto'))) as any
-    this._client = new protoDefinition.mesg.grpc.runner.Runner(options.endpoint || process.env.MESG_ENDPOINT, grpc.credentials.createInsecure())
-    this.credential = this.register(options.payload || process.env.MESG_REGISTER_PAYLOAD)
+    this._client = this.register(
+      options.signature || process.env.MESG_REGISTER_SIGNATURE,
+      options.endpoint || process.env.MESG_ENDPOINT,
+      process.env.MESG_SERVICE_HASH,
+      process.env.MESG_ENV_HASH
+    )
   }
 
-  async register(payload: string): Promise<grpc.Metadata> {
-    const res = await this.unaryCall('Register', { payload })
-    const meta = new grpc.Metadata()
-    meta.add('mesg_credential_token', res.token)
-    return meta
+  async register(signature: string, endpoint: string, serviceHash: string, envHash: string): Promise<grpc.GrpcObject> {
+    const runner = new Runner(endpoint)
+    const { token } = await runner.register({ serviceHash, envHash }, signature)
+    this._token = new grpc.Metadata()
+    this._token.add('mesg_credential_token', token)
+    const { mesg } = grpc.loadPackageDefinition(protoLoader.loadSync(path.join(__dirname, 'runner', 'runner.proto'), {
+      includeDirs: [__dirname]
+    })) as any
+    return new mesg.grpc.runner.Runner(endpoint, grpc.credentials.createInsecure())
   }
 
   listenTask({ ...tasks }: Tasks): EventEmitter {
@@ -41,8 +50,8 @@ class Service {
     }
     this.tasks = tasks;
     this.validateTaskNames();
-    this.credential.then(token => {
-      const stream = this._client.Execution({}, token) as grpc.ClientWritableStream<any>
+    this._client.then(client => {
+      const stream = client.Execution({}, this._token) as grpc.ClientWritableStream<any>
       stream.on('data', x => res.emit('data', x))
       stream.on('error', x => res.emit('error', x))
       stream.on('close', () => res.emit('close'))
@@ -57,7 +66,7 @@ class Service {
     return this.unaryCall('Event', {
       key: event,
       data: encode(data)
-    }, await this.credential)
+    }, this._token)
   }
 
   private async handleTaskData({ hash, taskKey, inputs }: IExecution) {
@@ -70,13 +79,13 @@ class Service {
       return this.unaryCall('Result', {
         executionHash: hash,
         outputs: encode(outputs)
-      }, await this.credential)
+      }, this._token)
     } catch (err) {
       const error = err.message;
       return this.unaryCall('Result', {
         executionHash: hash,
         error
-      }, await this.credential)
+      }, this._token)
     }
   }
 
@@ -91,13 +100,10 @@ class Service {
     }
   }
 
-  private async unaryCall(method: string, arg: any, credentials?: grpc.Metadata): Promise<any> {
+  private async unaryCall(method: string, arg: any, metadata: grpc.Metadata): Promise<any> {
+    const client = await this._client
     return new Promise((resolve, reject) => {
-      if (credentials) {
-        this._client[method](arg, credentials, (err: Error, res: any) => err ? reject(err) : resolve(res))
-      } else {
-        this._client[method](arg, (err: Error, res: any) => err ? reject(err) : resolve(res))
-      }
+      client[method](arg, metadata, (err: Error, res: any) => err ? reject(err) : resolve(res))
     })
   }
 }