+++ packages/api/CHANGELOG.md
@@ -18,6 +18,7 @@
 - [#207](https://github.com/mesg-foundation/js-sdk/pull/207) Update messages for the api and add integration tests
 - [#218](https://github.com/mesg-foundation/js-sdk/pull/218) Update filtering of transaction logs based on the new engine's events
 - [#220](https://github.com/mesg-foundation/js-sdk/pull/220) Expose crypto functions `Transaction.sign` and `Account.deriveMnemonic`
+- [#221](https://github.com/mesg-foundation/js-sdk/pull/221) Add struct type
 
 #### Bug fixes