+++ packages/service/CHANGELOG.md
@@ -6,6 +6,7 @@
 #### Improvements
 
 - [#216](https://github.com/mesg-foundation/js-sdk/pull/216) Add authentication to services
+- [#221](https://github.com/mesg-foundation/js-sdk/pull/221) Update registration of the service based on orchestrator API + specific runner client
 
 #### Bug fixes